refactor: restructure internal namespaces
This commit structures the API in the following way:
- the `penna.api.config` namespace should only contain interfaces;
- the `penna.api.models` will contain records/classes that are shared;
- any concrete implementation will be hidden in
  `penna.api.config.internal`.
