+++ penna-core/src/test/java/penna/core/logger/ConfigUpdateTests.java
@@ -3,7 +3,7 @@ package penna.core.logger;
 import org.junit.jupiter.api.Assertions;
 import org.junitpioneer.jupiter.cartesian.CartesianTest;
 import org.slf4j.event.Level;
-import penna.api.config.Config;
+import penna.api.models.Config;
 import penna.api.config.Manager;
 import penna.core.logger.guard.*;
 import penna.core.slf4j.PennaLoggerFactory;
@@ -14,8 +14,7 @@ public class ConfigUpdateTests {
     void canSetTheConfig(@CartesianTest.Enum Level level) {
         PennaLoggerFactory factory = PennaLoggerFactory.getInstance();
         PennaLogger logger = (PennaLogger) factory.getLogger(ConfigUpdateTests.class.getName());
-        Manager.Factory.initialize(factory);
-        Manager manager = Manager.Factory.getInstance();
+        Manager manager = Manager.create(factory);
         manager.set(ConfigUpdateTests.class.getName(), () -> Config.getDefault().replaceLevel(level));
         var target = switch (level) {
             case ERROR -> ErrorLevelGuard.class;
@@ -31,8 +30,7 @@ public class ConfigUpdateTests {
     void canUpdateTheConfig(@CartesianTest.Enum Level level) {
         PennaLoggerFactory factory = PennaLoggerFactory.getInstance();
         PennaLogger logger = (PennaLogger) factory.getLogger(ConfigUpdateTests.class.getName());
-        Manager.Factory.initialize(factory);
-        Manager manager = Manager.Factory.getInstance();
+        Manager manager = Manager.create(factory);
         manager.update(ConfigUpdateTests.class.getName(), (base) -> base.replaceLevel(level));
         var target = switch (level) {
             case ERROR -> ErrorLevelGuard.class;
